refactor: VoucherIoProcessor 유틸 패키지로 분류
