+++ src/main/java/study/dev/spring/voucher/presentation/ConsoleVoucherController.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import study.dev.spring.voucher.application.VoucherService;
 import study.dev.spring.voucher.application.dto.CreateVoucherRequest;
 import study.dev.spring.voucher.application.dto.VoucherInfo;
+import study.dev.spring.voucher.presentation.utils.VoucherIoProcessor;
 
 @Component
 @RequiredArgsConstructor
