+++ src/test/java/study/dev/spring/voucher/stub/VoucherIoProcessorStub.java
@@ -5,7 +5,7 @@ import java.util.List;
 import study.dev.spring.voucher.application.dto.CreateVoucherRequest;
 import study.dev.spring.voucher.application.dto.VoucherInfo;
 import study.dev.spring.voucher.fixture.VoucherFixture;
-import study.dev.spring.voucher.presentation.VoucherIoProcessor;
+import study.dev.spring.voucher.presentation.utils.VoucherIoProcessor;
 
 public class VoucherIoProcessorStub extends VoucherIoProcessor {
